// src/API.ts
export type Dish = {
  __typename: "Dish",
  id: string,
  name: string,
  description?: string | null,
  price: number,
  salePrice: number,
  images?: Array< string | null > | null,
  restaurantID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};